requirements.txt
python-dotenv
Pandas
sqlalchemy
pymysql